+++ cynic-codegen/src/query_dsl/interface_struct.rs
@@ -1,6 +1,8 @@
 use proc_macro2::TokenStream;
 
-use crate::{schema, Ident};
+use crate::{schema, type_index::TypeIndex, Ident};
+
+use super::SelectorStruct;
 
 /// We generate an InterfaceStruct for each interface in the schema.
 ///
@@ -9,12 +11,14 @@ use crate::{schema, Ident};
 #[derive(Debug)]
 pub struct InterfaceStruct {
     pub name: Ident,
+    pub selector_struct: SelectorStruct,
 }
 
 impl InterfaceStruct {
-    pub fn from_interface(interface: &schema::InterfaceType) -> Self {
+    pub fn from_interface(interface: &schema::InterfaceType, type_index: &TypeIndex) -> Self {
         InterfaceStruct {
             name: Ident::for_type(&interface.name),
+            selector_struct: SelectorStruct::new(&interface.name, &interface.fields, type_index),
         }
     }
 }
@@ -23,11 +27,8 @@ impl quote::ToTokens for InterfaceStruct {
     fn to_tokens(&self, tokens: &mut TokenStream) {
         use quote::{quote, TokenStreamExt};
 
-        let name = &self.name;
+        let selector = &self.selector_struct;
 
-        tokens.append_all(quote! {
-            #[allow(dead_code)]
-            pub struct #name {}
-        });
+        tokens.append_all(quote! { #selector });
     }
 }